# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creditra API
  version: 0.1.0
  description: |
    Backend API for Creditra – on-chain credit scoring and risk evaluation.

    Browser clients are restricted by a CORS allowlist. Production deployments
    must set `CORS_ORIGINS` to a comma-separated list of exact origins; local
    development falls back to loopback origins when the variable is unset.

    ## Authentication
    Protected endpoints require a valid API key passed via the `X-API-Key` header.
    API keys are configured via the `API_KEYS` environment variable (comma-separated).

    ## Keeping the spec in sync
    - **Zod schemas** in `src/schemas/` are the runtime source of truth for request
      and response validation. This OpenAPI document should mirror those schemas.
    - Every new route **must** have a corresponding entry here before merging.
    - Run `npm run validate:spec` in CI to catch YAML/structural errors early.
    - Use path-level `operationId` values as the source of truth for client generation.
    - See `docs/json-schema-validation.md` for request/response validation policy.
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authenticated endpoints. Required for admin operations.
        Keys are configured via the `API_KEYS` environment variable.
  schemas:
...
    CreateCreditLineRequest:
      type: object
      required: [walletAddress, requestedLimit]
      properties:
        walletAddress:
          type: string
          minLength: 1
          maxLength: 256
          description: Stellar wallet address
          example: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
        requestedLimit:
          type: string
          pattern: "^\\d+(\\.\\d+)?$"
          description: Requested credit limit as a decimal string
          example: "1000.00"
      additionalProperties: false
    UpdateCreditLineRequest:
      type: object
      description: |
        Partial update body (updateCreditLineSchema). At least one of the
        updatable fields must be present. Unknown keys are rejected.
      properties:
        creditLimit:
          type: string
          pattern: "^\\d+(\\.\\d+)?$"
          description: Updated credit limit
          example: "1500.00"
        interestRateBps:
          type: number
          description: Interest rate in basis points
          example: 640
        status:
          type: string
          enum: [active, suspended, closed]
          description: Credit line status
        expectedVersion:
          type: integer
          minimum: 1
          description: >
            Optimistic-locking guard. When provided, the update only succeeds
            if it equals the stored `version`; otherwise the server returns
            `409 Conflict`. Omit for unconditional (last-write-wins) updates.
          example: 3
      additionalProperties: false
    CreditLine:
      type: object
      properties:
        id:
          type: string
          description: Unique credit line identifier
          example: "cl_abc123"
        walletAddress:
          type: string
          description: Associated Stellar wallet address
          example: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
        creditLimit:
          type: string
          description: Approved credit limit
          example: "720.00000000"
        currentBalance:
          type: string
          description: Current outstanding balance
          example: "100.00000000"
        availableCredit:
          type: string
          description: Available credit (creditLimit - currentBalance)
          example: "620.00000000"
        interestRateBps:
          type: number
          description: Interest rate in basis points
          example: 640
        status:
          type: string
          enum: [active, suspended, closed]
          description: Current credit line status
        version:
          type: integer
          minimum: 1
          description: >
            Optimistic-locking version, incremented on each update. Pass the
            value you read back as `expectedVersion` on a subsequent update to
            guard against lost writes.
          example: 1
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last update timestamp
    CreditLinesResponse:
      type: object
      required: [creditLines, pagination]
      properties:
        creditLines:
          type: array
          items:
            $ref: "#/components/schemas/CreditLine"
        pagination:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
          required: [total, offset, limit]
    WalletCreditLinesResponse:
      type: object
      required: [creditLines]
      properties:
        creditLines:
          type: array
          items:
            $ref: "#/components/schemas/CreditLine"
    CreditLineActionResponse:
      type: object
      required: [line, message]
      properties:
        line:
          $ref: "#/components/schemas/CreditLine"
        message:
          type: string
          example: "Credit line suspended."

    # ── Transaction Schemas ──────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, creditLineId, type, amount, currency, timestamp, metadata]
      properties:
        id:
          type: string
          format: uuid
          description: Unique transaction identifier
        creditLineId:
          type: string
          description: ID of the credit line this transaction belongs to
        type:
          type: string
          enum: [borrow, repay, interest_accrual, fee, status_change]
          description: Transaction type
        amount:
          type: string
          nullable: true
          description: Transaction amount as a decimal string; null for status changes
          example: "500.00000000"
        currency:
          type: string
          nullable: true
          description: Currency code (e.g. USDC); null for status changes
          example: USDC
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the transaction occurred
        metadata:
          type: object
          additionalProperties: true
          description: Additional context; status changes include an 'action' field
          example:
            action: created
    TransactionHistoryResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [transactions, total, page, limit, totalPages]
          properties:
            transactions:
              type: array
              items:
                $ref: "#/components/schemas/Transaction"
            total:
              type: integer
              description: Total number of transactions matching the applied filters
            page:
              type: integer
              description: Current page number (1-indexed)
            limit:
              type: integer
              description: Maximum number of transactions per page
            totalPages:
              type: integer
              description: Total number of pages available

    # ── Risk Evaluation Schemas ───────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
          description: Always null on error
        error:
          type: string
        id:
          type: string
          description: Present on 404 responses that reference a specific resource
    ValidationErrorResponse:
      type: object
      required: [data, error, details]
      description: |
        Returned for Zod request-schema failures (body, query, or path params).
        Stable shape — never includes stack traces or internal exception text.
      properties:
        data:
          nullable: true
          example: null
        error:
          type: string
          enum: [Validation failed]
          example: Validation failed
        details:
          type: array
          items:
            type: object
            required: [field, message]
            additionalProperties: false
            properties:
              field:
                type: string
                description: Dotted path to the invalid field, or "(root)"
                example: walletAddress
              message:
                type: string
                example: walletAddress must be a valid Stellar address
      additionalProperties: false
    RateLimitErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Too many requests"
paths:
...
  /api/credit/lines:
    get:
      tags: [Credit]
      operationId: listCreditLines
      summary: List all credit lines
      security: []
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset (offset mode; ignored when `cursor` is present)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          description: Number of credit lines per page (validated by creditLinesQuerySchema)
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            When present (including empty string), switches to cursor pagination.
            Pass `nextCursor` from the previous page for subsequent pages.
      responses:
        "200":
          description: |
            Offset mode returns `{ data: { creditLines, pagination }, error: null }`.
            Cursor mode returns an unenveloped `{ creditLines, pagination }` body
            (see docs/cursor-pagination.md).
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed per window
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              description: Requests remaining in current window
              schema:
                type: integer
                example: 95
            X-RateLimit-Reset:
              description: Unix timestamp when the rate limit window resets
              schema:
                type: integer
                example: 1700000060
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLinesResponse"
              example:
                data:
                  creditLines: []
                  pagination:
                    total: 0
                    offset: 0
                    limit: 100
                error: null
        "400":
          description: Invalid query parameter (Zod validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
                example: 47
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitErrorResponse"

    post:
      tags: [Credit]
      operationId: createCreditLine
      summary: Create a new credit line
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditLineRequest"
            example:
              walletAddress: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
              requestedLimit: "1000.00"
      responses:
        "201":
          description: Credit line created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
        "400":
          description: Validation error or missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credit/lines/{id}:
    get:
      tags: [Credit]
      operationId: getCreditLineById
      summary: Get a credit line by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Credit line identifier
      responses:
        "200":
          description: Credit line found
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
        "404":
          description: Credit line not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Credit line not found
                id: "abc123"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitErrorResponse"

    put:
      tags: [Credit]
      operationId: updateCreditLine
      summary: Update a credit line
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Credit line identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreditLineRequest"
            example:
              creditLimit: "1500.00"
              interestRateBps: 640
              status: active
              expectedVersion: 3
      responses:
        "200":
          description: Credit line updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Credit line not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            Optimistic-locking conflict. The supplied `expectedVersion` did not
            match the stored `version`, meaning a concurrent update won the
            race. Re-read the credit line to get the current `version` and
            retry the update with it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Credit]
      operationId: deleteCreditLine
      summary: Delete a credit line
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Credit line identifier
      responses:
        "204":
          description: Credit line deleted successfully
        "404":
          description: Credit line not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credit/lines/{id}/transactions:
    get:
      tags: [Credit]
      operationId: getCreditLineTransactions
      summary: List transaction history for a credit line
      description: |
        Returns a paginated list of transactions for the given credit line.
        Supports optional filtering by transaction type and date range.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Credit line identifier
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [borrow, repay, interest_accrual, fee, status_change]
          description: Filter by transaction type
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include only transactions at or after this ISO 8601 timestamp (inclusive)
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include only transactions at or before this ISO 8601 timestamp (inclusive)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-indexed)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of transactions per page
      responses:
        "200":
          description: Paginated list of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHistoryResponse"
              example:
                data:
                  transactions:
                    - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                      creditLineId: "line-abc"
                      type: status_change
                      amount: null
                      currency: null
                      timestamp: "2024-01-15T10:00:00.000Z"
                      metadata:
                        action: created
                  total: 1
                  page: 1
                  limit: 20
                  totalPages: 1
        "400":
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid type filter. Must be one of: borrow, repay, interest_accrual, fee, status_change."
        "404":
          description: Credit line not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credit/lines/{id}/suspend:
    post:
      tags: [Credit]
      operationId: suspendCreditLine
      summary: Suspend a credit line
      description: |
        Suspends the specified credit line, preventing further draws.
        Requires API key authentication.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Credit line identifier
      responses:
        "200":
          description: Credit line suspended successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLineActionResponse"
        "401":
          description: API key missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Unauthorized
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
        "404":
          description: Credit line not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credit/lines/{id}/close:
    post:
      tags: [Credit]
      operationId: closeCreditLine
      summary: Close a credit line
      description: |
        Closes the specified credit line permanently.
        Requires API key authentication.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Credit line identifier
      responses:
        "200":
          description: Credit line closed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLineActionResponse"
        "401":
          description: API key missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Unauthorized
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
        "404":
          description: Credit line not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credit/wallet/{walletAddress}/lines:
    get:
      tags: [Credit]
      operationId: getCreditLinesByWallet
      summary: Get credit lines for a specific wallet
      security: []
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
          description: Stellar wallet address
      responses:
        "200":
          description: Credit lines for the wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletCreditLinesResponse"
              example:
                creditLines:
                  - id: "cl_abc123"
                    walletAddress: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
                    creditLimit: "720.00000000"
                    currentBalance: "100.00000000"
                    availableCredit: "620.00000000"
                    interestRateBps: 640
                    status: active
                    createdAt: "2024-01-15T10:00:00.000Z"
                    updatedAt: "2024-01-15T10:00:00.000Z"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Risk Evaluation Routes ──────────────────────────────────────────────────
...